// webapp/src/lib/api/model/memberRole.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Kakeibo API
 * マルチテナント家計簿システム API
 * OpenAPI spec version: 1.0.0
 */

export type MemberRole = (typeof MemberRole)[keyof typeof MemberRole];

export const MemberRole = {
  owner: "owner",
  editor: "editor",
  viewer: "viewer",
} as const;